web/requirements.txt
aniso8601==9.0.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
numpy==1.20.3
pandas==1.2.4
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.15
urllib3==1.26.4
Werkzeug==2.0.1
